+++ src/routes/index.mjs
@@ -1,0 +1,10 @@
+import { Router } from "express";
+import userRouter from "./user.mjs";
+import productsRouter from "./products.mjs";
+
+const router = Router();
+
+router.use(userRouter);
+router.use(productsRouter);
+
+export default router;
